requirements.txt
streamlit
langchain
openai
langchain_community
tiktoken
Pillow
python-dotenv
PyPDF2
